% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{ball_cover}
\alias{ball_cover}
\title{Ball cover}
\usage{
ball_cover(x, centers, radii, compress = FALSE)
}
\arguments{
\item{x}{point cloud, as a (n x d) matrix representing \code{n} points in \code{d} dimensions.}

\item{centers}{ball centers, as a (j x d) matrix representing \code{j} points in \code{d} dimensions.}

\item{radii}{either the radius of each ball, or a single radius for all the balls}
}
\description{
Constructs a ball cover over a point set at a given radius.
}
\details{
This function determines the set of points that lie within the union of closed balls
of some supplied \code{radius}.
}
