% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverr-package.R
\docType{package}
\name{coverr-package}
\alias{coverr}
\alias{coverr-package}
\title{coverr: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
