% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{interval_cover}
\alias{interval_cover}
\title{interval_cover}
\usage{
interval_cover(x, num_intervals, overlap, bounded = FALSE)
}
\arguments{
\item{x}{point cloud to construct a cover of.}

\item{num_intervals}{number of intervals in the cover, per dimension.}

\item{overlap}{proportion of overlap between intervals.}
}
\description{
This function constructs an interval over on the unit hypercube \link{0,1}^d.
}
