% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{knn_cover}
\alias{knn_cover}
\title{K-Nearest Neighbor cover}
\usage{
knn_cover(x, k, n_sets = 25L)
}
\description{
This function determines the set of points that lie within the union of the top \code{n_sets}
densest balls according to a \code{k}-NN density estimate. The radii of the these balls is proportional
to distance to their k-th nearest neighbors.
}
